+++ posthog/hogql_queries/insights/funnels/test/conversion_time_cases.py
@@ -7,7 +7,7 @@ from posthog.schema import FunnelsQuery
 
 from posthog.constants import INSIGHT_FUNNELS, FunnelOrderType
 from posthog.hogql_queries.insights.funnels.funnels_query_runner import FunnelsQueryRunner
-from posthog.hogql_queries.insights.funnels.test.test_funnel import PseudoFunnelActors
+from posthog.hogql_queries.insights.funnels.test.test_utils import PseudoFunnelActors
 from posthog.hogql_queries.legacy_compatibility.filter_to_query import filter_to_query
 from posthog.models.filters import Filter
 from posthog.test.test_journeys import journeys_for
